Separate implementation and definitions for the *Sample classes.
